rename caseflow_only, refactor into concern
