+++ app/models/board_grant_effectuation.rb
@@ -2,6 +2,8 @@
 # Some are represented as contentions on an EP in VBMS. Others are tracked via Caseflow tasks.
 
 class BoardGrantEffectuation < ApplicationRecord
+  include Benefitable
+
   belongs_to :appeal
   belongs_to :granted_decision_issue, class_name: "DecisionIssue"
   belongs_to :decision_document
@@ -27,10 +29,6 @@ class BoardGrantEffectuation < ApplicationRecord
     granted_decision_issue.benefit_type
   end
 
-  def effectuated_in_vbms?
-    ClaimantValidator::BENEFIT_TYPE_REQUIRES_PAYEE_CODE.include?(benefit_type)
-  end
-
   def hydrate_from_granted_decision_issue
     assign_attributes(
       appeal: granted_decision_issue.decision_review,
@@ -42,12 +40,6 @@ class BoardGrantEffectuation < ApplicationRecord
     else
       find_or_build_effectuation_task
     end
-  end
-
-  # TODO: Refactor with ClaimReview business_line in a concern
-  def business_line
-    business_line_name = Constants::BENEFIT_TYPES[benefit_type]
-    @business_line ||= BusinessLine.find_or_create_by(url: benefit_type, name: business_line_name)
   end
 
   def find_or_build_effectuation_task
