+++ app/models/decision_review_intake.rb
@@ -5,7 +5,7 @@ class DecisionReviewIntake < Intake
       claimant: detail.claimant_participant_id,
       veteran_is_not_claimant: detail.veteran_is_not_claimant,
       payeeCode: detail.payee_code,
-      nonComp: detail.caseflow_only?,
+      nonComp: !detail.benefit_type_requires_payee_code?,
       legacy_opt_in_approved: detail.legacy_opt_in_approved,
       legacyAppeals: detail.serialized_legacy_appeals,
       ratings: detail.serialized_ratings,
